Build Graph
Initialize and empty object.
Create a modified roads array, where each road is split into an array of two elements: [start, end]
Loop over the array and repeatedly add key-value pairs (as per the expected output) to an object you had initialized earlier.
During each iteration, decide what happens based on if that place is already in the object or not.
A ternary operator may be of use in this decision making process.